реализовал переключение сортировок через StatefulView
