+++ src/mock/points.js
@@ -24,49 +24,6 @@ const OFFERS = [
   'Aliquam erat volutpat.'
 ];
 
-const SORTINGS = [
-  {
-    classList: 'trip-sort__item--day',
-    value: 'sort-day',
-    id: 'sort-day',
-    label: 'Day',
-    disabled: false,
-    sortType: SortType.DAY,
-  },
-  {
-    classList: 'trip-sort__item--event',
-    value: 'sort-event',
-    id: 'sort-event',
-    label: 'Event',
-    disabled: true,
-    sortType: SortType.EVENT,
-  },
-  {
-    classList: 'trip-sort__item--time',
-    value: 'sort-time',
-    id: 'sort-time',
-    label: 'Time',
-    disabled: false,
-    sortType: SortType.TIME,
-  },
-  {
-    classList: 'trip-sort__item--price',
-    value: 'sort-price',
-    id: 'sort-price',
-    label: 'Price',
-    disabled: false,
-    sortType: SortType.PRICE,
-  },
-  {
-    classList: 'trip-sort__item--offer',
-    value: 'sort-offer',
-    id: 'sort-offer',
-    label: 'Offers',
-    disabled: false,
-    sortType: SortType.OFFER,
-  },
-];
-
 const createDestination = () => (
   {
     'id': nanoid(),
@@ -142,6 +99,5 @@ const points = Array.from({length: 5}, createPoint);
 export const getPoints = () => points;
 export const getOffers = () => offers;
 export const getDestinations = () => destinations;
-export const getSortings = () => SORTINGS;
 export const getOffersByType = () => offersByType;
